JDF-344 Fix for JavaScript error during subsequent bookings.
Contains multiple fixes to ensure that the ticket shopping cart in the mobile view behaves almost similarly as the desktop one.

Fixed the back button behavior in checkout. The back button now navigates to the create booking view.

The ticket summary view has also been changed to use a listview instead of table. This is to allow bookings to be deleted via the split button design, as well as to aid in responsive design. Booking can be deleted at the level of categories.

Also fixed an issue involving a JavaScript exception thrown on triggering the jQueryMobile pagecreate event in the Create Booking view.